Refactor: split file in two
