+++ src/main/kotlin/com/gmaur/investment/robotadvisor/infrastructure/DomainObjectMapper.kt
@@ -6,23 +6,6 @@ import java.math.BigDecimal
 import java.math.MathContext
 import java.math.RoundingMode
 
-
-data class AssetAllocationElementDTO(val isin: String, val percentage: String)
-data class AssetAllocationDTO(val listOf: List<AssetAllocationElementDTO>)
-data class PortfolioDTO(val values: List<XDTO>)
-data class XDTO(val asset: AssetDTO, val amount: AmountDTO)
-
-data class OperationsDTO(val operations: List<OperationDTO>)
-data class OperationDTO(val type: String, val asset: AssetDTO, val amount: AmountDTO)
-data class AssetDTO(val isin: String, val transferrable: Boolean)
-data class AmountDTO private constructor(val value: String, val currency: String) {
-    companion object {
-        fun EUR(value: String): AmountDTO {
-            return AmountDTO(value, "EUR")
-        }
-    }
-}
-
 class DomainObjectMapper {
     fun toDTO(it: Operations): OperationsDTO {
         return OperationsDTO(
